DEV: Update create-invite modal to component-based API
